Fixed sonar

+++ src/main/java/g3401_3500/s3430_maximum_and_minimum_sums_of_at_most_size_k_subarrays/Solution.java
@@ -37,7 +37,7 @@ public class Solution {
                 ans += (long) num * l * r;
             } else {
                 long cnt = (long) (k - r + 1) * r + (long) (l + r - k - 1) * (r + k - l) / 2;
-                ans += (long) num * cnt;
+                ans += num * cnt;
             }
         }
         return ans;
